requirements.txt
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
polars==1.27.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
